provider-registry/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
asyncpg==0.29.0
psycopg[binary]==3.1.19
alembic==1.12.1
pydantic==2.5.0
opentelemetry-instrumentation-fastapi==0.42b0
opentelemetry-exporter-otlp==1.21.0
opentelemetry-sdk==1.21.0
prometheus-fastapi-instrumentator==6.1.0